+++ R/timss99.R
@@ -1,0 +1,6 @@
+#' Random TIMSS 99 data with 3000 cases and 3 countries
+#'
+#' @name timss99
+#' @docType data
+#' @keywords data
+NULL